+++ crates/client/devnet/src/lib.rs
@@ -196,7 +196,6 @@ mod tests {
     use mp_block::header::L1DataAvailabilityMode;
     use mp_block::{BlockId, BlockTag};
     use mp_class::ClassInfo;
-    use mp_convert::felt_to_u128;
     use mp_receipt::{Event, ExecutionResult, FeePayment, InvokeTransactionReceipt, PriceUnit, TransactionReceipt};
     use mp_transactions::compute_hash::calculate_contract_address;
     use mp_transactions::BroadcastedTransactionExt;
@@ -300,7 +299,7 @@ mod tests {
     }
 
     fn chain_with_mempool_limits(mempool_limits: MempoolLimits) -> DevnetForTesting {
-        let _ = env_logger::builder().is_test(true).try_init();
+        let _ = tracing_subscriber::fmt().with_test_writer().try_init();
 
         let mut g = ChainGenesisDescription::base_config().unwrap();
         let contracts = g.add_devnet_contracts(10).unwrap();
@@ -586,7 +585,7 @@ mod tests {
             false => {
                 assert_eq!(&receipt.execution_result, &ExecutionResult::Succeeded);
 
-                let fees_fri = felt_to_u128(&block.inner.receipts[0].actual_fee().amount).unwrap();
+                let fees_fri = block.inner.receipts[0].actual_fee().amount.try_into().unwrap();
                 assert_eq!(chain.get_bal_strk_eth(sequencer_address), (fees_fri, 0));
                 assert_eq!(
                     chain.get_bal_strk_eth(contract_0.address),
@@ -601,7 +600,7 @@ mod tests {
                 let ExecutionResult::Reverted { reason } = receipt.execution_result else { unreachable!() };
                 assert!(reason.contains("ERC20: insufficient balance"));
 
-                let fees_fri = felt_to_u128(&block.inner.receipts[0].actual_fee().amount).unwrap();
+                let fees_fri = block.inner.receipts[0].actual_fee().amount.try_into().unwrap();
                 assert_eq!(chain.get_bal_strk_eth(sequencer_address), (fees_fri, 0));
                 assert_eq!(
                     chain.get_bal_strk_eth(contract_0.address),