+++ crates/client/rpc/src/versions/v0_7_1/methods/read/block_hash_and_number.rs
@@ -34,7 +34,7 @@ mod tests {
     };
     use mp_state_update::StateDiff;
     use rstest::rstest;
-    use starknet_core::types::Felt;
+    use starknet_types_core::felt::Felt;
 
     #[rstest]
     fn test_block_hash_and_number(rpc_test_setup: (Arc<MadaraBackend>, Starknet)) {