+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 - feat(block_production): continue pending block on restart
 - feat(mempool): mempool transaction saving on db
 - feat(mempool): mempool transaction limits
+- refactor: replace starknet-rs BlockId with types-rs BlockId and remove redundant mp_block::BlockId
+- feat(fgw): added `add_transaction` for gateway client
+- fix(fgw): include `l1_to_l2_consumed_message` in L1 handler receipt
+- build: up starknet-rs, starknet-types, blockifier(v0.8.0), cairo
+- feat(rpc): added `getCompiledCasm` method
+- fix(error): Added a comment for non archive node L1 keys
 - feat(confg): added chain config template and fgw example
 - feat(v0.8.0-rc0): starknet_subscribeNewHeads
 - fix(rocksdb): update max open files opt